Upgrade to SDK 35 and update build configuration
- Updated Android Gradle Plugin from 7.4.2 to 8.7.2
- Updated Gradle from 7.6 to 8.9
- Updated compileSdkVersion and targetSdkVersion to 35
- Updated buildToolsVersion to 35.0.0
- Updated Java compatibility from 1.8 to 17
- Added namespace declaration (required for AGP 8.x)
- Enabled buildFeatures.buildConfig
- Updated packaging and lint configuration syntax
- Fixed resource ID switch statements for AGP 8.x compatibility
- Updated Moshi to use reflection-based adapters instead of kapt
- Fixed Snackbar text view reference
- Updated Kotlin toShort() conversions for Java 17 compatibility
- Increased JVM memory allocation

+++ app/src/main/java/org/tyndalebt/storyproduceradv/tools/media/pipe/PipedAudioMixer.kt
@@ -135,7 +135,7 @@ class PipedAudioMixer : PipedAudioShortManipulator(), PipedMediaByteBufferDest {
             val mb = mixBuffers[iSource].sliceArray(pos..pos+allLength)
 
             //setup the data as a source.  Copy the buffer
-            srcBuffer.forEachIndexed { index, sh ->  srcBuffer[index] = (sh + mb[index]).toShort()}
+            srcBuffer.forEachIndexed { index, sh ->  srcBuffer[index] = (sh + mb[index]).toInt().toShort()}
 
             mixPoss[iSource] += allLength
             iSource++
